add students and lecturers to courses

+++ src/main/java/com/ftn/studentservice/repository/LectureInstanceRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LectureInstanceRepository extends JpaRepository<LectureInstance, Integer> {
+public interface LectureInstanceRepository extends JpaRepository<LectureInstance, Long> {
 
 
 }
