+++ src/main/java/com/ftn/studentservice/repository/LecturerRepository.java
@@ -7,10 +7,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LecturerRepository extends JpaRepository<Lecturer, Integer> {
+public interface LecturerRepository extends JpaRepository<Lecturer, Long> {
 
    public Page<Lecturer> findByLectureInstances_CourseInstance_Id(Long id, Pageable pageable);
    public Page<Lecturer> findAllByLecturerCodeContainsOrUser_FirstNameContainsOrUser_LastNameContains(String index,String firstName,String lastName, Pageable pageable);
    public Lecturer findByUser_Id(Long id);
+
+
+
 
 }
